router: Change user modify module transfer method
For the user modify view, we used Id(as string) in UserFormat to update
the data before. But it didn't work.

Therefore, we use the Hex string of bson.ObejectID representing
Id of userformat. Use Hex() and ObjectIDHex() to convert the _id of
mongodb and general ID of storage formate in golang.

As a result, this commit mainly updates the convert process between
mongodb operation and the transmitted accountID in URL.